adding Action Types
